New settings pertaining to webseeds. These settings can be changed within the settings.json file.
Note that none of these specific settings are
described on the official transmission web site.
The advanced webseed algorithm used
in the releases here has fixed most,
if not all, of the bugs posted on the bug trac system
of the official transmission website.
Official ticket https://trac.transmissionbt.com/ticket/5923
summarizes many of the problems, and offers fixes,
however this release uses at times a quite different,
but very well tested, approach to solve the problems
presented there.