Fixed: https://github.com/transmission/transmission/issues/129 Fixed: "Unable to save resume file: File name too long" for magnet links with extremely long names. Fixed: "Invalid or corrupt torrent file" for torrents with an empty name key.
http://bittorrent.org/beps/bep_0003.html

20-Oct-2012 (arvid@bittorrent.com), clarified that info-hash is the digest of en bencoding found in .torrent file.

The 20 byte sha1 hash of the bencoded form of the info value from the metainfo file. Note that this is a substring of the metainfo file. The info-hash must be the hash of the encoded form as found in the .torrent file, regardless of it being invalid.

The name key maps to a UTF-8 encoded string which is the suggested name to save the file (or directory) as. It is purely advisory.