Further fixes for https://github.com/transmission/transmission/issues/129
https://github.com/bittorrent/bittorrent.org/commit/1a5fd50896cd85dd46f65ed6033981868131fd92
"They must not perform a decode-encode roundtrip on invalid data."

FIXED: Incoming magnet links, with redundant keys, are at first hashed correctly. Then tr_bencMergeDicts() invokes various '* see if it already exists, and if so, try to reuse it *' and tr_bencListCopy() is only called for nonexistent keys. The result was a .torrent file not exact info dictionary of the incoming info dict, and different (incorrect) hash for incoming magnet at tr_metainfoParseImpl().